+++ libs/libmod/src/mod/lib/RC/ComposeRuleRealGeneric.hpp
@@ -75,6 +75,23 @@ struct MatchMakerCallback {
 		using HasTerm = GraphMorphism::HasTermData<InvertibleVertexMap>;
 		using HasStereo = GraphMorphism::HasStereoData<InvertibleVertexMap>;
 		constexpr LabelType labelType = HasTerm::value ? LabelType::Term : LabelType::String;
+		if(verbose) {
+			logger.indent() << "RuleComp: " << rFirst.getName() << " with " << rSecond.getName() << std::endl;
+			logger.indent() << "   hasTerm=" << std::boolalpha << HasTerm::value << ", hasStereo=" << HasStereo::value << std::endl;
+			logger.indent() << "   map:";
+			const auto &gFirst = get_R_projected(rFirst.getDPORule());
+			const auto &gSecond = get_L_projected(rSecond.getDPORule());
+			bool first = true;
+			for(const auto vSecond: asRange(vertices(gSecond))) {
+				const auto vFirst = get(m, gSecond, gFirst, vSecond);
+				if(vFirst == gFirst.null_vertex()) continue;
+				if(first) logger.s << " ";
+				else logger.s << ", ";
+				logger.s << "(" << get(boost::vertex_index_t(), gFirst, vFirst) << ", "
+				         << get(boost::vertex_index_t(), gSecond, vSecond) << ")";
+			}
+			logger.s << std::endl;
+		}
 		auto rResult = composeRuleRealByMatch<labelType, HasStereo::value>(rFirst, rSecond, m, verbose, logger);
 		if(rResult) {
 			if(verbose)
@@ -86,6 +103,11 @@ struct MatchMakerCallback {
 			}
 			const bool cont = rr(std::move(rResult));
 			if(!cont) return false;
+		} else {
+			if(verbose)
+				logger.indent() << "RuleComp failed"
+				                << "\t " << rFirst.getName()
+				                << "\t. " << rSecond.getName() << std::endl;
 		}
 		return true;
 	}